Patch svelte config's base path for gh-pages
GitHub Pages is hosting this repo at "/<REPO>/" but svelte was
configured to use "/". This caused no other files, except for
index.html to be resolved properly. Fix this by patching the config
prior to running gh-pages.

I've chosen to add a patch file and apply the patch on each deployment
instead of modifying the svelte config directly since I'd prefer to be
able to deploy/host this project anywhere. To do so, I just need to add
a new patch file that adheres to the new hosts environment. Adding
these rules to the svelte config would get cluttered and not apply most
of the time.