Remove SvelteKit, TailwindCSS, and npm
While this project was rather "complete", it felt like it was
overwhelming to understand; and it is just a single page static site!
I decided to move back to vanilla HTML5, CSS3, and JavaScript.

This commit will remove all of the old files and move to a simpler
project structure. Given the project's simple and narrow scope, this
should provide an easier understanding for everyone when looking at the
source. For more complex projects, I do see the advantages of
SvelteKit, TailwindCSS, and npm. But lets focus on the basics here!